# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Leptech LMS API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API server"
        }
    ],
    "paths": {
        "/admin/sign-in": {
            "post": {
                "tags": [
                    "Admin Authentication"
                ],
                "operationId": "90d67c15e6d11bcf4bd4400225c8da01",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Logged in"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    },
                    "403": {
                        "description": "Not allowed in admin panel"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/admin/sign-out": {
            "post": {
                "tags": [
                    "Admin Authentication"
                ],
                "operationId": "f545716a49630865ae51174555a5179e",
                "responses": {
                    "200": {
                        "description": "Logged out"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/org/create": {
            "post": {
                "tags": [
                    "Organization"
                ],
                "operationId": "7d333c555114e616c17d6f759a599de3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "shortDescription"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Acme Inc."
                                    },
                                    "shortDescription": {
                                        "type": "string",
                                        "example": "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Maecenas venenatis enim at dolor rutrum, non volutpat metus faucibus. Aliquam tincidunt lorem non nisi vestibulum fermentum."
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Aenean orci velit, varius vitae massa pellentesque, dapibus rhoncus dui. Proin at fringilla ipsum, non venenatis erat. Mauris cursus elit vel dictum condimentum. Mauris rhoncus auctor convallis. Mauris nisl tellus, aliquam lobortis interdum sed, consectetur vitae nulla. Integer laoreet metus nec neque scelerisque, sed lacinia diam interdum. Sed vitae ante nulla. Quisque eu felis nec orci molestie blandit. Pellentesque ac enim vestibulum, molestie nisi eget, interdum metus. Suspendisse sagittis auctor nibh vel sollicitudin. Sed dignissim urna eu est egestas iaculis. Nulla nec porta massa. Duis nunc elit, ornare sit amet justo quis."
                                    },
                                    "noOfUsers": {
                                        "type": "integer",
                                        "example": 10,
                                        "default": 0
                                    },
                                    "noOfAdmins": {
                                        "type": "integer",
                                        "example": 2,
                                        "default": 0
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Organization created"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Not allowed"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "409": {
                        "description": "Organization already exists"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/org/all": {
            "get": {
                "tags": [
                    "Organization"
                ],
                "operationId": "824979088ed76f53dfa6c630a148597d",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "maximum": 100
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of organizations"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Not allowed"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/org/{organization}": {
            "patch": {
                "tags": [
                    "Organization"
                ],
                "operationId": "2ed1caa9160f2b75bffe5862c9e5d225",
                "parameters": [
                    {
                        "name": "organization",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Acme Inc."
                                    },
                                    "shortDescription": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "noOfUsers": {
                                        "type": "integer"
                                    },
                                    "noOfAdmins": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Organization updated"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Not allowed"
                    },
                    "404": {
                        "description": "Organization not found"
                    },
                    "409": {
                        "description": "Organization name already exists"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/test": {
            "get": {
                "tags": [
                    "Test"
                ],
                "operationId": "2cf1648f26e6c36635ef5a3b5321fcff",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Paste the token returned by /admin/login",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Admin Authentication",
            "description": "Admin Authentication"
        },
        {
            "name": "Organization",
            "description": "Organization"
        },
        {
            "name": "Test",
            "description": "Test"
        }
    ]
}